대댓글 테이블 제거, 댓글 테이블 self-join 으로 변경, 댓글 리스트 개발 ing

+++ board/src/main/kotlin/com/unimal/board/service/post/dto/PostInfo.kt
@@ -17,7 +17,7 @@ data class PostInfo(
     val createdAt: LocalDateTime,
     val fileInfoList: List<BoardFileInfo?>,
     val likeCount: Long,
-    val replyCount: Int,
+    val replyCount: Long,
     val reply: List<Reply>,
     val isOwner: Boolean = false
 )
